cell_track/piprequirements_gnn.txt
absl-py==0.15.0
aicsimageio==4.13.0
aicspylibczi==3.1.2
aiobotocore==2.5.4
aiohttp==3.8.1
aioitertools==0.11.0
aiosignal==1.2.0
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.4.0
backcall==0.2.0
blinker==1.4
botocore==1.31.17
Bottleneck==1.3.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
colorlog==6.8.0
comm==0.2.2
commonmark==0.9.1
cryptography==36.0.0
cycler==0.11.0
cytoolz==0.11.0
dask==2022.2.1
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.2
distributed==2022.2.1
docker-pycreds==0.4.0
elementpath==4.4.0
entrypoints==0.4
executing==2.0.1
faiss==1.7.2
fasteners==0.19
filelock==3.6.0
flatbuffers==20210226132247
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2023.6.0
future==0.18.2
gast==0.3.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==1.35.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googledrivedownloader==0.4
greenlet==3.0.3
grpcio==1.37.1
h5py==2.10.0
hydra-colorlog==1.0.0
hydra-core==1.1.0.dev5
hydra-optuna-sweeper==0.9.0rc2
identify==2.5.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
importlib-metadata==4.11.3
importlib-resources==6.1.1
iniconfig==1.1.1
ipykernel==6.29.3
ipython==8.12.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter_core==5.7.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
locket==0.2.1
lxml==4.9.4
Mako==1.3.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
msgpack==1.0.8
multidict==5.2.0
munkres==1.1.4
nest_asyncio==1.6.0
networkx==2.7.1
nodeenv==1.6.0
numcodecs==0.12.1
numexpr==2.7.3
numpy==1.22.3
nvidia-ml-py==12.535.133
nvitop==1.3.2
oauthlib==3.1.0
ome-types==0.5.1.post1
ome-zarr==0.8.3
omegaconf==2.1.0.dev24
opt-einsum==3.3.0
optuna==3.5.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
partd==1.2.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
platformdirs==4.2.0
pluggy==1.0.0
pre-commit==2.19.0
prompt-toolkit==3.0.42
protobuf==3.15.8
psutil==5.9.8
ptyprocess==0.7.0
pudb==2022.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.7.0
pydantic-compat==0.1.2
pydantic_core==2.18.1
pyDeprecate==0.3.1
Pygments==2.11.2
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.4
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-louvain==0.15
pytorch-lightning==1.4.9
pytorch-metric-learning==1.3.0
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.1.0
record-keeper==0.9.32
requests==2.27.1
requests-oauthlib==1.3.0
resource-backed-dask-array==0.1.0
rich==10.16.2
rsa==4.7.2
s3fs==2023.6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.5.3
seaborn==0.11.2
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==61.2.0
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
stack-data==0.6.2
tblib==3.0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2023.2.28
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.8.0
torch-cluster==1.5.9
torch-geometric==2.0.1
torch-scatter==2.0.8
torch-sparse==0.6.18
torch-spline-conv==1.2.1
torchmetrics==0.6.0
torchvision==0.9.0a0
tornado==6.1
tqdm==4.64.0
traitlets==5.14.2
typing_extensions==4.9.0
ukkonen==1.0.1
urllib3==1.26.18
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.4.6
wandb==0.16.2
wcwidth==0.2.13
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.13.3
xarray==2023.1.0
xmlschema==3.3.0
xsdata==24.3.1
yacs==0.1.6
yarl==1.6.3
zarr==2.15.0
zict==3.0.0
zipp==3.7.0
